projects/datastore/pyproject.toml
[project]
name = "datastore"
version = "0.1.0"
description = "Datastore microservice for mechadealer project"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "python-dotenv>=1.1.0",
    "questdb[dataframe]>=2.0.3",
    "upstox-python-sdk>=2.14.0",
    "pandas>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["datastore"]  # Updated to point directly to datastore directory
